20201214リファクタリング
