+++ 06_DB操作系/TSAパスワード管理/settings.json
@@ -1,0 +1,4 @@
+{
+    "log_filename" : "./log/manageTsaPW.log",
+    "db_file" : "ng_passwd.db"
+}
